+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -252,9 +252,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-08-25T10:39:50.840Z"
+    created: Optional[str] = "2026-08-25T12:34:46.571Z"
 
-    updated: Optional[str] = "2026-08-25T10:39:50.840Z"
+    updated: Optional[str] = "2026-08-25T12:34:46.571Z"
 
     updated_by_id: OptionalNullable[str] = UNSET