+++ packages/orq-rc/RELEASES.md
@@ -10459,3 +10459,13 @@ Based on:
 - [python v4.14.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v4.14.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.52 - packages/orq-rc
+
+## 2026-08-25 12:38:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.3 (2.932.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.14.0-rc.53] packages/orq-rc
+### Releases
+- [PyPI v4.14.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.53 - packages/orq-rc